add single phase meter support, error handling
Single phase meter will update the total column. Error handling added to not parse json on http (tested with 500) errors. Try to detect single or three phase meters, to reduce the number of (unnecessary) API calls for single phase meters. Small improvement to not color trends on first measurement. Amps shown with a fixed number of decimal digits (5 A -> 5.0 A)